Revert "Support flattened data_stream.* fields (#3465)"
This reverts commit b0b8e85bce33642de86f2f13a81b3544b3f8e778.
