--- changelog/fragments/1695389490-Support-flattened-data_stream.-fields.yaml
@@ -1,35 +1,0 @@
-# Kind can be one of:
-# - breaking-change: a change to previously-documented behavior
-# - deprecation: functionality that is being removed in a later release
-# - bug-fix: fixes a problem in a previous version
-# - enhancement: extends functionality but does not break or fix existing behavior
-# - feature: new functionality
-# - known-issue: problems that we are aware of in a given version
-# - security: impacts on the security of a product or a user’s deployment.
-# - upgrade: important information for someone upgrading from a prior version
-# - other: does not fit into any of the other categories
-kind: feature
-
-# Change summary; a 80ish characters long description of the change.
-summary: Support flattened data_stream.* fields in input configuration
-
-# Long description; in case the summary is not enough to describe the change
-# this field accommodate a description without length limits.
-# NOTE: This field will be rendered only for breaking-change and known-issue kinds at the moment.
-description: >-
-  An input configuration supports flattened fields, however the
-  'data_stream' field was not being correctly decoded when
-  flattened. This commit fixes this issue.
-
-# Affected component; usually one of "elastic-agent", "fleet-server", "filebeat", "metricbeat", "auditbeat", "all", etc.
-component: elastic-agent
-
-# PR URL; optional; the PR number that added the changeset.
-# If not present is automatically filled by the tooling finding the PR where this changelog fragment has been added.
-# NOTE: the tooling supports backports, so it's able to fill the original PR number instead of the backport PR number.
-# Please provide it if you are adding a fragment for a different PR.
-pr: https://github.com/elastic/elastic-agent/pull/3465
-
-# Issue URL; optional; the GitHub issue related to this changeset (either closes or is part of).
-# If not present is automatically filled by the tooling with the issue linked to the PR number.
-issue: https://github.com/elastic/elastic-agent/issues/3191
