+++ pkg/component/component_test.go
@@ -2361,74 +2361,3 @@ func gatherDurationFieldPaths(s interface{}, pathSoFar string) []string {
 
 	return gatheredPaths
 }
-
-func TestFlattenedDataStream(t *testing.T) {
-	expectedNamespace := "test-namespace"
-	expectedType := "test-type"
-	expectedDataset := "test-dataset"
-
-	policy := map[string]any{
-		"outputs": map[string]any{
-			"default": map[string]any{
-				"type":    "elasticsearch",
-				"enabled": true,
-			},
-		},
-		"inputs": []any{
-			map[string]any{
-				"type":                "filestream",
-				"id":                  "filestream-0",
-				"enabled":             true,
-				"data_stream.type":    expectedType,
-				"data_stream.dataset": expectedDataset,
-				"data_stream": map[string]any{
-					"namespace": expectedNamespace,
-				},
-			},
-		},
-	}
-	runtime, err := LoadRuntimeSpecs(filepath.Join("..", "..", "specs"), PlatformDetail{}, SkipBinaryCheck())
-	if err != nil {
-		t.Fatalf("cannot load runtime specs: %s", err)
-	}
-
-	result, err := runtime.ToComponents(policy, nil, logp.DebugLevel, nil)
-	if err != nil {
-		t.Fatalf("cannot convert policy to component: %s", err)
-	}
-
-	if len(result) != 1 {
-		t.Fatalf("expecting result to have one element, got %d", len(result))
-	}
-
-	if len(result[0].Units) != 2 {
-		t.Fatalf("expecting result[0].Units to have two elements, got %d", len(result))
-	}
-
-	// We do not make assumptions about ordering.
-	// Get the input Unit
-	var dataStream *proto.DataStream
-	for _, unit := range result[0].Units {
-		if unit.Err != nil {
-			t.Fatalf("unit.Err: %s", unit.Err)
-		}
-		if unit.Type == client.UnitTypeInput {
-			dataStream = unit.Config.DataStream
-			break
-		}
-	}
-
-	if dataStream == nil {
-		t.Fatal("DataStream cannot be nil")
-	}
-
-	if dataStream.Dataset != expectedDataset {
-		t.Errorf("expecting DataStream.Dataset: %q, got: %q", expectedDataset, dataStream.Dataset)
-	}
-	if dataStream.Type != expectedType {
-		t.Errorf("expecting DataStream.Type: %q, got: %q", expectedType, dataStream.Type)
-	}
-	if dataStream.Namespace != expectedNamespace {
-		t.Errorf("expecting DataStream.Namespace: %q, got: %q", expectedNamespace, dataStream.Namespace)
-	}
-}
