+++ pkg/testing/fixture.go
@@ -357,14 +357,7 @@ func (f *Fixture) RunBeat(ctx context.Context) error {
 // Elastic Agent is stopped. If at any time the Elastic Agent logs an error log and the Fixture is not started
 // with `WithAllowErrors()` then `Run` will exit early and return the logged error.
 //
-// If no `states` are provided then the Elastic Agent runs until the context is cancelled.
-//
-// The Elastic-Agent is started agent in test mode (--testing-mode) this mode
-// expects the initial configuration (full YAML config) via gRPC.
-// This configuration should be passed in the State.Configure field.
-//
-// The `elastic-agent.yml` generated by `Fixture.Configure` is ignored
-// when `Run` is called.
+// If no `states` are provided then the Elastic Agent runs until the context is or the timeout specified with WithRunLength is reached.
 func (f *Fixture) Run(ctx context.Context, states ...State) error {
 	if f.binaryName != "elastic-agent" {
 		return errors.New("Run() can only be used with elastic-agent, use RunBeat()")
